refactor replay worker
